version2 kimi_k2/requirements.txt
torch>=2.0
transformers
pandas
numpy
scikit-learn
yfinance
ta
finnhub-python
hydra-core
omegaconf
loguru
fastapi
uvicorn
pytest